# perf: improve speed of parsing