Hello!

I rewrote the entire function to avoid creating a class during parsing and also calling methods, here are the results of that rewrite:

```
parseBoolArray('{true}') x 3,600,315 ops/sec ±1.18% (89 runs sampled)
parseBoolArray('{true,false}') x 1,504,719 ops/sec ±1.34% (87 runs sampled)
parseBoolArray('{true,false,true}') x 1,074,796 ops/sec ±1.27% (88 runs sampled)
parseBoolArrayV2('{true}') x 8,257,642 ops/sec ±1.46% (87 runs sampled)
parseBoolArrayV2('{true,false}') x 4,202,999 ops/sec ±1.93% (88 runs sampled)
parseBoolArrayV2('{true,false,true}') x 3,031,024 ops/sec ±1.51% (86 runs sampled)
Fastest is parseBoolArrayV2('{true}')
```

> [Benchmark](https://gist.github.com/H4ad/2600b4b8067e0a52b642a733c357c125)

It improves parsing speed by about 2x to 3x times.